Added a track and a note implementation
It is now possible to play a simple song. A Note now has an
implementation. There are plans to give this more functionality,
however there are things that are more urgent. Also the SimpleTrack
is quite badly written. However for small tracks it should work.
Plans are made for making a better faster implementation of it.